refactor(wasm): split sensor inputs + install/snapshot out of lib.rs
Two more #[wasm_bindgen] impl blocks split from the WasmSimulator god-impl:
- wasm/inputs.rs: sensor / board-IO injectors (HC-SR04, I2C sensors, ADC,
  NTC/MAX31855, GPS, SN74HC165 — 14 methods).
- wasm/install.rs: firmware-quirk installers + runtime-snapshot save/restore
  (6 methods).

lib.rs 1659 -> 538 (2540 -> 538 across the wasm PRs) — now just the struct,
constructors, and core execution. Verified wasm32 + host builds, clippy, fmt.
